Update patch set 1
Patch Set 1:

This was only tested and confirmed on Doubleshot.  Owner of commit has a Doubleshot.  Considering the similarities with Pyramid (our kernel base, our device repo, etc), this change might work for Pyramid or other HTC MSM8660 devices too.  We did not confirm.

With Doubleshot, bootup is noticeably faster with this change.

Patch-set: 1